requirements.txt
google-generativeai
textblob
pandas
python-dotenv
